+++ src/viat/_cli/commands/rm.py
@@ -13,12 +13,11 @@ from viat.vault import autoload_vault
 def rm(ctx: click.Context, path: pathlib.Path) -> None:
     """Remove a file along with its metadata."""
     vault = autoload_vault(ctx.obj.vault_config)
-    rel_path = vault.normalize_path(path)
 
     try:
         path.unlink()
     except OSError as err:
         raise ViatVaultError(f'Aborting due to file system error: {err}') from err
 
-    with vault.storage as conn, conn.get_mutator(rel_path) as mut:
+    with vault.storage as conn, conn.get_mutator(path) as mut:
         mut.clear()