+++ src/viat/_cli/commands/update.py
@@ -25,9 +25,9 @@ def update(ctx: click.Context, path: pathlib.Path, attrs: str) -> None:
         raise ViatMalformedDataError(f'Expected a JSON object, but got {attrs!r}')
 
     vault = autoload_vault(ctx.obj.vault_config)
-    rel_path = vault.normalize_path(path)
+    vault.tracker.validate_tracked(path)
 
-    with vault.storage as conn, conn.get_mutator(rel_path) as mut:
+    with vault.storage as conn, conn.get_mutator(path) as mut:
         mut.update(parsed)
         updated = dict(mut)